hardware_web_server/requirements.txt
APScheduler==3.8.1
async-timeout==4.0.2
attrs==21.4.0
backports.zoneinfo==0.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
coverage==6.2
DBUtils==2.0.1
Flask==1.1.2
Flask-APScheduler==1.12.2
flask-blueprint==1.3.0
Flask-Cors==3.0.10
Flask-HTTPAuth==4.2.0
gevent==21.1.2
gevent-websocket==0.10.1
greenlet==1.1.0
gunicorn==20.1.0
idna==2.10
importlib-metadata==4.8.3
importlib-resources==5.4.0
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
packaging==21.3
pika==1.3.1
pluggy==1.0.0
py==1.11.0
PyMySQL==1.0.2
pyncclient==0.7
pyparsing==3.0.7
pytest==7.0.1
pytest-cov==3.0.0
python-dateutil==2.8.2
python-dotenv==0.17.1
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
redis==4.3.5
requests==2.25.1
six==1.15.0
tomli==1.2.3
typing_extensions==4.1.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.4
Werkzeug==1.0.1
zipp==3.6.0
zope.event==4.5.0
zope.interface==5.4.0
